Ouroboros.Consensus.ByronDual.Ledger

Shorthand

type DualByronBlock = DualBlock ByronBlock ByronSpecBlock

type DualByronBridge = BridgeLedger ByronBlock ByronSpecBlock

Bridge

data ByronSpecBridge

Bridge the gap between the Byron implementation and specification

The relation between the Byron implementation and specification for the linear case is tested in the Byron implementation itself, specifically in ts_prop_generatedChainsAreValidated. The main goal of the consensus DualByron tests is to lift these tests to the general consensus setting, where time is not linear but branching.

In the linear case, the tests maintain some state linking the spec and the implementation. In the consensus case, this state cannot be maintained like this, and so it has to become part of transactions, blocks, and the ledger state itself.

Constructors

ByronSpecBridge

Fields

toSpecKeys ∷ Map (PBftVerKeyHash PBftByronCrypto) VKey

Map between keys

Some observations:

The abstract chain environment contains a set of allowed delegators (of type Set VKeyGenesis), which gets translated to gdGenesisKeyHashes (of type Set Common.KeyHash) in the concrete genesis config.

During the translation from abstract blocks to concrete blocks, the VKey of the block is translated to a concrete SigningKey (as well as a VerificationKey) in elaborateKeyPair.

Although this translation is deterministic, it doesn't have an easily definable inverse. For this reason, we maintain an opposite mapping as part of the ledger state.

toImplIds ∷ SpecToImplIds

Mapping between abstract and concrete Ids

We need to maintain this mapping so that we can use the abstract state generators and then elaborate to concrete values.

newtype SpecToImplIds

Constructors

SpecToImplIds

Fields

getSpecToImplIds ∷ AbstractToConcreteIdMaps

bridgeToSpecKey ∷ DualByronBridge → PBftVerKeyHash PBftByronCrypto → VKey

Translate issuer key

We get a proof from PBFT that we are the leader, including a signing key (of type SigningKey). In order to produce the corresponding abstract block, we need a VKey.

bridgeTransactionIds ∷ DualByronBridge → Map TxId TxId

initByronSpecBridge

Arguments

∷ ByronSpecGenesis

→ Map TxId TxId
Mapping for the transaction in the initial UTxO

→ ByronSpecBridge

specToImplTx ∷ Tx → ATxAux ByteString → SpecToImplIds

Construct singleton SpecToImplIds for a transaction

Block forging

forgeDualByronBlock

Arguments

∷ HasCallStack

⇒ TopLevelConfig DualByronBlock

→ BlockNo
Current block number

→ SlotNo
Current slot number

→ TickedLedgerState DualByronBlock
Ledger

→ [Validated (GenTx DualByronBlock)]
Txs to add in the block

→ PBftIsLeader PBftByronCrypto
Leader proof (IsLeader)

→ DualByronBlock

Orphan instances

Bridge ByronBlock ByronSpecBlock

Instance details

Associated Types

type BridgeLedger ByronBlock ByronSpecBlock

type BridgeBlock ByronBlock ByronSpecBlock

type BridgeTx ByronBlock ByronSpecBlock

Methods

updateBridgeWithBlock ∷ DualBlock ByronBlock ByronSpecBlock → BridgeLedger ByronBlock ByronSpecBlock → BridgeLedger ByronBlock ByronSpecBlock

updateBridgeWithTx ∷ Validated (GenTx (DualBlock ByronBlock ByronSpecBlock)) → BridgeLedger ByronBlock ByronSpecBlock → BridgeLedger ByronBlock ByronSpecBlock