Ouroboros.Consensus.ByronDual.Node.Serialisation

Orphan instances

SerialiseNodeToNodeConstraints DualByronBlock

Instance details

Methods

estimateBlockSize ∷ Header DualByronBlock → SizeInBytes

SerialiseNodeToClientConstraints DualByronBlock

Instance details

SerialiseDiskConstraints DualByronBlock

Instance details

HasNetworkProtocolVersion DualByronBlock

Instance details

Associated Types

type BlockNodeToNodeVersion DualByronBlock

type BlockNodeToClientVersion DualByronBlock

SupportedNetworkProtocolVersion DualByronBlock

Instance details

Methods

supportedNodeToNodeVersions ∷ Proxy DualByronBlock → Map NodeToNodeVersion (BlockNodeToNodeVersion DualByronBlock)

supportedNodeToClientVersions ∷ Proxy DualByronBlock → Map NodeToClientVersion (BlockNodeToClientVersion DualByronBlock)

latestReleasedNodeVersion ∷ Proxy DualByronBlock → (Maybe NodeToNodeVersion, Maybe NodeToClientVersion)

SerialiseNodeToNode DualByronBlock DualByronBlock
CBOR-in-CBOR for the annotation. This also makes it compatible with the wrapped (Serialised) variant.

Instance details

Methods

encodeNodeToNode ∷ CodecConfig DualByronBlock → BlockNodeToNodeVersion DualByronBlock → DualByronBlock → Encoding

decodeNodeToNode ∷ CodecConfig DualByronBlock → BlockNodeToNodeVersion DualByronBlock → ∀ s. Decoder s DualByronBlock

SerialiseNodeToClient DualByronBlock SlotNo

Instance details

Methods

encodeNodeToClient ∷ CodecConfig DualByronBlock → BlockNodeToClientVersion DualByronBlock → SlotNo → Encoding

decodeNodeToClient ∷ CodecConfig DualByronBlock → BlockNodeToClientVersion DualByronBlock → ∀ s. Decoder s SlotNo

SerialiseNodeToClient DualByronBlock DualByronBlock
CBOR-in-CBOR for the annotation. This also makes it compatible with the wrapped (Serialised) variant.

Instance details

Methods

encodeNodeToClient ∷ CodecConfig DualByronBlock → BlockNodeToClientVersion DualByronBlock → DualByronBlock → Encoding

decodeNodeToClient ∷ CodecConfig DualByronBlock → BlockNodeToClientVersion DualByronBlock → ∀ s. Decoder s DualByronBlock

EncodeDisk DualByronBlock DualByronBlock

Instance details

Methods

encodeDisk ∷ CodecConfig DualByronBlock → DualByronBlock → Encoding

SerialiseNodeToNode DualByronBlock (Serialised DualByronBlock)
CBOR-in-CBOR for the annotation. This also makes it compatible with the wrapped (Serialised) variant.

Instance details

Methods

encodeNodeToNode ∷ CodecConfig DualByronBlock → BlockNodeToNodeVersion DualByronBlock → Serialised DualByronBlock → Encoding

decodeNodeToNode ∷ CodecConfig DualByronBlock → BlockNodeToNodeVersion DualByronBlock → ∀ s. Decoder s (Serialised DualByronBlock)

SerialiseNodeToNode DualByronBlock (SerialisedHeader DualByronBlock)
Forward to the Byron instance.

Instance details

Methods

encodeNodeToNode ∷ CodecConfig DualByronBlock → BlockNodeToNodeVersion DualByronBlock → SerialisedHeader DualByronBlock → Encoding

decodeNodeToNode ∷ CodecConfig DualByronBlock → BlockNodeToNodeVersion DualByronBlock → ∀ s. Decoder s (SerialisedHeader DualByronBlock)

SerialiseNodeToNode DualByronBlock (GenTx DualByronBlock)

Instance details

Methods

encodeNodeToNode ∷ CodecConfig DualByronBlock → BlockNodeToNodeVersion DualByronBlock → GenTx DualByronBlock → Encoding

decodeNodeToNode ∷ CodecConfig DualByronBlock → BlockNodeToNodeVersion DualByronBlock → ∀ s. Decoder s (GenTx DualByronBlock)

SerialiseNodeToNode DualByronBlock (GenTxId DualByronBlock)

Instance details

Methods

encodeNodeToNode ∷ CodecConfig DualByronBlock → BlockNodeToNodeVersion DualByronBlock → GenTxId DualByronBlock → Encoding

decodeNodeToNode ∷ CodecConfig DualByronBlock → BlockNodeToNodeVersion DualByronBlock → ∀ s. Decoder s (GenTxId DualByronBlock)

SerialiseNodeToNode DualByronBlock (Header DualByronBlock)
Forward to the Byron instance.

Instance details

Methods

encodeNodeToNode ∷ CodecConfig DualByronBlock → BlockNodeToNodeVersion DualByronBlock → Header DualByronBlock → Encoding

decodeNodeToNode ∷ CodecConfig DualByronBlock → BlockNodeToNodeVersion DualByronBlock → ∀ s. Decoder s (Header DualByronBlock)

SerialiseNodeToClient DualByronBlock (Serialised DualByronBlock)
CBOR-in-CBOR for the annotation. This also makes it compatible with the wrapped (Serialised) variant.

Instance details

Methods

encodeNodeToClient ∷ CodecConfig DualByronBlock → BlockNodeToClientVersion DualByronBlock → Serialised DualByronBlock → Encoding

decodeNodeToClient ∷ CodecConfig DualByronBlock → BlockNodeToClientVersion DualByronBlock → ∀ s. Decoder s (Serialised DualByronBlock)

SerialiseNodeToClient DualByronBlock (GenTx DualByronBlock)

Instance details

Methods

encodeNodeToClient ∷ CodecConfig DualByronBlock → BlockNodeToClientVersion DualByronBlock → GenTx DualByronBlock → Encoding

decodeNodeToClient ∷ CodecConfig DualByronBlock → BlockNodeToClientVersion DualByronBlock → ∀ s. Decoder s (GenTx DualByronBlock)

SerialiseNodeToClient DualByronBlock (GenTxId DualByronBlock)

Instance details

Methods

encodeNodeToClient ∷ CodecConfig DualByronBlock → BlockNodeToClientVersion DualByronBlock → GenTxId DualByronBlock → Encoding

decodeNodeToClient ∷ CodecConfig DualByronBlock → BlockNodeToClientVersion DualByronBlock → ∀ s. Decoder s (GenTxId DualByronBlock)

SerialiseResult DualByronBlock (BlockQuery DualByronBlock)

Instance details

Methods

encodeResult ∷ CodecConfig DualByronBlock → BlockNodeToClientVersion DualByronBlock → BlockQuery DualByronBlock result → result → Encoding

decodeResult ∷ CodecConfig DualByronBlock → BlockNodeToClientVersion DualByronBlock → BlockQuery DualByronBlock result → ∀ s. Decoder s result

EncodeDisk DualByronBlock (PBftState PBftByronCrypto)
ChainDepState (BlockProtocol DualByronBlock)

Instance details

Methods

encodeDisk ∷ CodecConfig DualByronBlock → PBftState PBftByronCrypto → Encoding

EncodeDisk DualByronBlock (AnnTip DualByronBlock)

Instance details

Methods

encodeDisk ∷ CodecConfig DualByronBlock → AnnTip DualByronBlock → Encoding

EncodeDisk DualByronBlock (LedgerState DualByronBlock)

Instance details

Methods

encodeDisk ∷ CodecConfig DualByronBlock → LedgerState DualByronBlock → Encoding

DecodeDisk DualByronBlock (PBftState PBftByronCrypto)
ChainDepState (BlockProtocol DualByronBlock)

Instance details

Methods

decodeDisk ∷ CodecConfig DualByronBlock → ∀ s. Decoder s (PBftState PBftByronCrypto)

DecodeDisk DualByronBlock (AnnTip DualByronBlock)

Instance details

Methods

decodeDisk ∷ CodecConfig DualByronBlock → ∀ s. Decoder s (AnnTip DualByronBlock)

DecodeDisk DualByronBlock (LedgerState DualByronBlock)

Instance details

Methods

decodeDisk ∷ CodecConfig DualByronBlock → ∀ s. Decoder s (LedgerState DualByronBlock)

SerialiseNodeToClient DualByronBlock (DualGenTxErr ByronBlock ByronSpecBlock)
ApplyTxErr DualByronBlock

Instance details

Methods

encodeNodeToClient ∷ CodecConfig DualByronBlock → BlockNodeToClientVersion DualByronBlock → DualGenTxErr ByronBlock ByronSpecBlock → Encoding

decodeNodeToClient ∷ CodecConfig DualByronBlock → BlockNodeToClientVersion DualByronBlock → ∀ s. Decoder s (DualGenTxErr ByronBlock ByronSpecBlock)

SerialiseNodeToClient DualByronBlock (SomeSecond BlockQuery DualByronBlock)

Instance details

Methods

encodeNodeToClient ∷ CodecConfig DualByronBlock → BlockNodeToClientVersion DualByronBlock → SomeSecond BlockQuery DualByronBlock → Encoding

decodeNodeToClient ∷ CodecConfig DualByronBlock → BlockNodeToClientVersion DualByronBlock → ∀ s. Decoder s (SomeSecond BlockQuery DualByronBlock)

DecodeDisk DualByronBlock (ByteString → DualByronBlock)

Instance details

Methods

decodeDisk ∷ CodecConfig DualByronBlock → ∀ s. Decoder s (ByteString → DualByronBlock)

DecodeDiskDep (NestedCtxt Header) DualByronBlock

Instance details

Methods

decodeDiskDep ∷ CodecConfig DualByronBlock → NestedCtxt Header DualByronBlock a → ∀ s. Decoder s (ByteString → a)