Test.Consensus.Byron.Examples

Setup

secParam ∷ SecurityParam

Note that we must use the same value for the SecurityParam as for the WindowSize, because decodeByronChainDepState only takes the SecurityParam and uses it as the basis for the WindowSize.

windowSize ∷ WindowSize

cfg ∷ BlockConfig ByronBlock

codecConfig ∷ CodecConfig ByronBlock

ledgerConfig ∷ LedgerConfig ByronBlock

leaderCredentials ∷ ByronLeaderCredentials

Examples

examples ∷ Examples ByronBlock

exampleChainDepState ∷ ChainDepState (BlockProtocol ByronBlock)

exampleLedgerState ∷ LedgerState ByronBlock

exampleHeaderState ∷ HeaderState ByronBlock

exampleExtLedgerState ∷ ExtLedgerState ByronBlock

exampleHeaderHash ∷ ByronHash

exampleGenTx ∷ GenTx ByronBlock

exampleGenTxId ∷ TxId (GenTx ByronBlock)

exampleApplyTxErr ∷ ApplyMempoolPayloadErr