Test.Consensus.Byron.Generators

Documentation

newtype RegularBlock

A ByronBlock that is never an EBB.

Constructors

RegularBlock

Fields

unRegularBlock ∷ ByronBlock

epochSlots ∷ EpochSlots

Matches that from the dummyConfig

k ∷ SecurityParam

Matches that from the dummyConfig

protocolMagicId ∷ ProtocolMagicId

Orphan instances

Arbitrary ProtocolParameters

Instance details

Methods

arbitrary ∷ Gen ProtocolParameters

shrink ∷ ProtocolParameters → [ProtocolParameters]

Arbitrary ByronTransition

Instance details

Methods

arbitrary ∷ Gen ByronTransition

shrink ∷ ByronTransition → [ByronTransition]

Arbitrary ByronNodeToNodeVersion

Instance details

Methods

arbitrary ∷ Gen ByronNodeToNodeVersion

shrink ∷ ByronNodeToNodeVersion → [ByronNodeToNodeVersion]

Arbitrary ByronNodeToClientVersion

Instance details

Methods

arbitrary ∷ Gen ByronNodeToClientVersion

shrink ∷ ByronNodeToClientVersion → [ByronNodeToClientVersion]

Arbitrary ByronHash

Instance details

Methods

arbitrary ∷ Gen ByronHash

shrink ∷ ByronHash → [ByronHash]

Arbitrary ByronBlock

Instance details

Methods

arbitrary ∷ Gen ByronBlock

shrink ∷ ByronBlock → [ByronBlock]

Arbitrary KeyHash

Instance details

Methods

arbitrary ∷ Gen KeyHash

shrink ∷ KeyHash → [KeyHash]

Arbitrary ChainValidationState

Instance details

Methods

arbitrary ∷ Gen ChainValidationState

shrink ∷ ChainValidationState → [ChainValidationState]

Arbitrary UTxO

Instance details

Methods

arbitrary ∷ Gen UTxO

shrink ∷ UTxO → [UTxO]

Arbitrary ApplyMempoolPayloadErr

Instance details

Methods

arbitrary ∷ Gen ApplyMempoolPayloadErr

shrink ∷ ApplyMempoolPayloadErr → [ApplyMempoolPayloadErr]

Arbitrary ProtocolVersion

Instance details

Methods

arbitrary ∷ Gen ProtocolVersion

shrink ∷ ProtocolVersion → [ProtocolVersion]

Arbitrary State

Instance details

Methods

arbitrary ∷ Gen State

shrink ∷ State → [State]

Arbitrary Map

Instance details

Methods

arbitrary ∷ Gen Map

shrink ∷ Map → [Map]

Arbitrary SoftwareVersion

Instance details

Methods

arbitrary ∷ Gen SoftwareVersion

shrink ∷ SoftwareVersion → [SoftwareVersion]

Arbitrary GenesisHash

Instance details

Methods

arbitrary ∷ Gen GenesisHash

shrink ∷ GenesisHash → [GenesisHash]

Arbitrary EpochNumber

Instance details

Methods

arbitrary ∷ Gen EpochNumber

shrink ∷ EpochNumber → [EpochNumber]

Arbitrary SlotNumber

Instance details

Methods

arbitrary ∷ Gen SlotNumber

shrink ∷ SlotNumber → [SlotNumber]

Arbitrary ApplicationName

Instance details

Methods

arbitrary ∷ Gen ApplicationName

shrink ∷ ApplicationName → [ApplicationName]

Arbitrary InstallerHash

Instance details

Methods

arbitrary ∷ Gen InstallerHash

shrink ∷ InstallerHash → [InstallerHash]

Arbitrary SystemTag

Instance details

Methods

arbitrary ∷ Gen SystemTag

shrink ∷ SystemTag → [SystemTag]

Arbitrary State

Instance details

Methods

arbitrary ∷ Gen State

shrink ∷ State → [State]

Arbitrary ScheduledDelegation

Instance details

Methods

arbitrary ∷ Gen ScheduledDelegation

shrink ∷ ScheduledDelegation → [ScheduledDelegation]

Arbitrary ApplicationVersion

Instance details

Methods

arbitrary ∷ Gen ApplicationVersion

shrink ∷ ApplicationVersion → [ApplicationVersion]

Arbitrary ProtocolUpdateProposal

Instance details

Methods

arbitrary ∷ Gen ProtocolUpdateProposal

shrink ∷ ProtocolUpdateProposal → [ProtocolUpdateProposal]

Arbitrary SoftwareUpdateProposal

Instance details

Methods

arbitrary ∷ Gen SoftwareUpdateProposal

shrink ∷ SoftwareUpdateProposal → [SoftwareUpdateProposal]

Arbitrary State

Instance details

Methods

arbitrary ∷ Gen State

shrink ∷ State → [State]

Arbitrary State

Instance details

Methods

arbitrary ∷ Gen State

shrink ∷ State → [State]

Arbitrary (PBftState PBftByronCrypto)

Instance details

Methods

arbitrary ∷ Gen (PBftState PBftByronCrypto)

shrink ∷ PBftState PBftByronCrypto → [PBftState PBftByronCrypto]

Arbitrary (AnnTip ByronBlock)

Instance details

Methods

arbitrary ∷ Gen (AnnTip ByronBlock)

shrink ∷ AnnTip ByronBlock → [AnnTip ByronBlock]

Arbitrary (TipInfoIsEBB ByronBlock)

Instance details

Methods

arbitrary ∷ Gen (TipInfoIsEBB ByronBlock)

shrink ∷ TipInfoIsEBB ByronBlock → [TipInfoIsEBB ByronBlock]

Arbitrary (GenTx ByronBlock)

Instance details

Methods

arbitrary ∷ Gen (GenTx ByronBlock)

shrink ∷ GenTx ByronBlock → [GenTx ByronBlock]

Arbitrary (GenTxId ByronBlock)

Instance details

Methods

arbitrary ∷ Gen (GenTxId ByronBlock)

shrink ∷ GenTxId ByronBlock → [GenTxId ByronBlock]

Arbitrary (LedgerState ByronBlock)

Instance details

Methods

arbitrary ∷ Gen (LedgerState ByronBlock)

shrink ∷ LedgerState ByronBlock → [LedgerState ByronBlock]

Arbitrary (Header ByronBlock)

Instance details

Methods

arbitrary ∷ Gen (Header ByronBlock)

shrink ∷ Header ByronBlock → [Header ByronBlock]

Arbitrary (Coherent ByronBlock)

Instance details

Methods

arbitrary ∷ Gen (Coherent ByronBlock)

shrink ∷ Coherent ByronBlock → [Coherent ByronBlock]

Arbitrary (SomeResult ByronBlock)

Instance details

Methods

arbitrary ∷ Gen (SomeResult ByronBlock)

shrink ∷ SomeResult ByronBlock → [SomeResult ByronBlock]

Arbitrary (Hash a)

Instance details

Methods

arbitrary ∷ Gen (Hash a)

shrink ∷ Hash a → [Hash a]

Arbitrary (SomeSecond BlockQuery ByronBlock)

Instance details

Methods

arbitrary ∷ Gen (SomeSecond BlockQuery ByronBlock)

shrink ∷ SomeSecond BlockQuery ByronBlock → [SomeSecond BlockQuery ByronBlock]

Arbitrary (WithVersion ByronNodeToNodeVersion (Header ByronBlock))
We only have to be careful about headers with ByronNodeToNodeVersion1, where we will have a fake block size hint.

Instance details

Methods

arbitrary ∷ Gen (WithVersion ByronNodeToNodeVersion (Header ByronBlock))

shrink ∷ WithVersion ByronNodeToNodeVersion (Header ByronBlock) → [WithVersion ByronNodeToNodeVersion (Header ByronBlock)]

Arbitrary (WithVersion ByronNodeToNodeVersion (SomeSecond (NestedCtxt Header) ByronBlock))

Instance details

Methods

arbitrary ∷ Gen (WithVersion ByronNodeToNodeVersion (SomeSecond (NestedCtxt Header) ByronBlock))

shrink ∷ WithVersion ByronNodeToNodeVersion (SomeSecond (NestedCtxt Header) ByronBlock) → [WithVersion ByronNodeToNodeVersion (SomeSecond (NestedCtxt Header) ByronBlock)]