Test.ThreadNet.Infra.Byron.ProtocolInfo

Documentation

mkLeaderCredentials ∷ HasCallStack ⇒ Config → GeneratedSecrets → CoreNodeId → ByronLeaderCredentials

mkProtocolByron

Arguments

∷ ∀ m. (Monad m, HasCallStack)

⇒ PBftParams

→ CoreNodeId

→ Config

→ GeneratedSecrets

→ (ProtocolInfo m ByronBlock, SignKeyDSIGN ByronDSIGN)
We return the signing key which is needed in some tests, because it cannot easily be extracted from the ProtocolInfo.

theProposedProtocolVersion ∷ ProtocolVersion

The protocol version proposed as part of the hard-fork smoke test

The initial Byron ledger state begins with protocol version 0.0.0. In the smoke test, if the proposal and votes are enabled, then we will be proposing an update to 1.0.0.

This value occurs in two crucial places: the proposal and also the byronProtocolVersion field of the static node config. See the Haddock comment on mkProtocolByronAndHardForkTxs.

theProposedSoftwareVersion ∷ SoftwareVersion

The software version proposed as part of the hard-fork smoke test

We don't actually care about this for the smoke test, but we have to set it both as part of the proposal and also as part of the node's static configuration. Its use in the static configuration is legacy and does not seem to affect anything; see Issue #1732.

The initial Byron ledger state begins with no recorded software versions. For the addition of a new software version, the Byron ledger rules require that it starts at 0 or 1.